+++ demo/index.js
@@ -2,7 +2,7 @@ import Vditor from '../src/index'
 import '../src/assets/scss/classic.scss'
 
 window.vditor = new Vditor('vditor', {
-  mode: 'wysiwyg-show',
+  mode: 'markdown-show',
   counter: 100,
   height: 300,
   hint: {
@@ -24,10 +24,7 @@ window.vditor = new Vditor('vditor', {
         replace(/[\?\\/:|<>\*\[\]\(\)\$%\{\}@~]/g, '').
         replace('/\\s/g', '')
     },
-  },
-  preview: {
-    mode: 'both',
-  },
+  }
 })
 
 window.vditorTest = new Vditor('vditorTest', {