+++ src/ts/util/Options.ts
@@ -31,7 +31,7 @@ export class Options {
             delay: 1000,
             hljs: {
                 enable: true,
-                style: "atom-one-light",
+                style: "github",
             },
             maxWidth: 768,
             mode: "both",